blank/pyproject.toml
[project]
name = "hud-blank"
version = "0.1.0"
description = "A minimal counter-based HUD environment for testing"
requires-python = ">=3.11"
dependencies = [
    "hud-python>=0.5.29",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "httpx>=0.24.0",
    "anthropic>=0.84.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

